Bug_900159 - Factor shared/js/device_storage_utils by creating separate files under shared/device_storage/ r=davidflanagan
