+++ apps/gallery/index.html
@@ -14,6 +14,7 @@
     <link rel="resource" type="application/l10n" href="locales/locales.ini" />
     <!-- Shared code -->
     <script defer src="shared/js/l10n.js"></script>
+    <script defer src="shared/js/device_storage/enumerate_all.js"></script>
     <script defer src="shared/js/mediadb.js"></script>
     <script defer src="shared/js/lazy_loader.js"></script>
     <script defer src="shared/js/visibility_monitor.js"></script>
