+++ apps/gallery/open.html
@@ -15,7 +15,8 @@
     <script type="text/javascript" defer src="shared/js/media/media_frame.js"></script>
     <script type="text/javascript" defer src="shared/js/blobview.js"></script>
     <script type="text/javascript" defer src="shared/js/media/jpeg_metadata_parser.js"></script>
-    <script type="text/javascript" defer src="shared/js/device_storage_utils.js"></script>
+    <script type="text/javascript" defer src="shared/js/device_storage/get_storage_if_available.js"></script>
+    <script type="text/javascript" defer src="shared/js/device_storage/get_unused_filename.js"></script>
     <!-- Specific code -->
     <script type="text/javascript" defer src="js/imagesize.js"></script>
     <script type="text/javascript" defer src="js/open.js"></script>
