+++ apps/music/open.html
@@ -12,7 +12,8 @@
     <link rel="resource" type="application/l10n" href="locales/locales.ini">
     <!-- Shared code -->
     <script type="text/javascript" src="shared/js/l10n.js"></script>
-    <script type="text/javascript" src="shared/js/device_storage_utils.js"></script>
+    <script type="text/javascript" src="shared/js/device_storage/get_storage_if_available.js"></script>
+    <script type="text/javascript" src="shared/js/device_storage/get_unused_filename.js"></script>
     <script type="text/javascript" defer src="shared/js/blobview.js"></script>
     <script type="text/javascript" defer src="shared/js/async_storage.js"></script>
     <!-- Specific code -->
