+++ apps/video/index.html
@@ -10,6 +10,7 @@
     <link rel="stylesheet" type="text/css" href="style/video.css" />
 
     <script defer src="shared/js/l10n.js"></script>
+    <script defer src="shared/js/device_storage/enumerate_all.js"></script>
     <script defer src="shared/js/mediadb.js"></script>
     <script defer src="shared/js/blobview.js"></script>
     <script defer src="shared/js/media/get_video_rotation.js"></script>
