+++ apps/video/view.html
@@ -47,7 +47,8 @@
     </section>
 
     <script type="text/javascript" src="shared/js/l10n.js"></script>
-    <script type="text/javascript" src="shared/js/device_storage_utils.js"></script>
+    <script type="text/javascript" src="shared/js/device_storage/get_storage_if_available.js"></script>
+    <script type="text/javascript" src="shared/js/device_storage/get_unused_filename.js"></script>
     <script defer src="shared/js/blobview.js"></script>
     <script defer src="shared/js/media/get_video_rotation.js"></script>
     <script type="text/javascript" src="js/view.js"></script>
