+++ test_apps/music2/index.html
@@ -70,7 +70,6 @@
       <!--==================  panels ================== -->
         <script src="js/UI/MediaLibrary/MediaLibraryMainPanel.js"></script>
         <script src="js/UI/MediaLibrary/MediaLibraryCategoryPanel.js"></script>
-
         <script src="js/UI/MediaLibrary/MediaLibraryPlaylistPanel.js"></script>
         <script src="js/UI/MediaLibrary/MediaLibraryPlaylistsPanel.js"></script>
         <script src="js/UI/MediaLibrary/MediaLibraryDiscoverPanel.js"></script>
@@ -87,6 +86,7 @@
         <script src="js/UI/MediaLibrary/MediaLibraryPageNotifications.js"></script>
 
   <!--==================  app ================== -->
+    <script src="shared/js/device_storage/enumerate_all.js"></script>
     <script src="shared/js/mediadb.js"></script>
     <script src="shared/js/async_storage.js"></script>
